+++ draft-packages/modal/KaizenDraft/Modal/Presets/ContextModal.tsx
@@ -42,8 +42,6 @@ export type ContextModalProps = Readonly<
   } & ContextModalSecondaryActionProps
 >
 
-type ContextModal = React.FunctionComponent<ContextModalProps>
-
 /**
  * {@link https://cultureamp.design/components/modal/#context-modals-previously-information-modal Guidance} |
  * {@link https://cultureamp.design/storybook/?path=/docs/components-modal--context-modal-example Storybook}
@@ -63,7 +61,7 @@ const ContextModal = ({
   contentHeader,
   image,
   ...props
-}: ContextModalProps) => {
+}: ContextModalProps): JSX.Element => {
   const onDismiss = confirmWorking ? undefined : props.onDismiss
 
   const footerActions: ButtonProps[] = []