+++ draft-packages/tile/KaizenDraft/Tile/components/Action.tsx
@@ -2,15 +2,17 @@ import React from "react"
 import { Button } from "@kaizen/button"
 import { TileAction } from "./GenericTile"
 
-interface Props {
+interface ActionProps {
   readonly action: TileAction
   readonly secondary?: boolean
   readonly disabled?: boolean
 }
 
-type Action = React.FunctionComponent<Props>
-
-const Action: Action = ({ action, secondary = false, disabled = false }) => {
+const Action = ({
+  action,
+  secondary = false,
+  disabled = false,
+}: ActionProps): JSX.Element => {
   const {
     label,
     href,