+++ packages/select/src/Select/Select.tsx
@@ -17,12 +17,6 @@ import { TriggerButton, TriggerButtonProps } from "./components/TriggerButton"
 import { SelectContext } from "./context/SelectContext"
 import selectStyles from "./Select.module.scss"
 
-type SubComponentProps = {
-  TriggerButton: typeof TriggerButton
-  Option: typeof Option
-  ListBox: typeof ListBox
-}
-
 export type SelectOptionsProps = {
   items: Array<Node<SingleItemType>>
 }
@@ -52,7 +46,7 @@ export interface SelectProps
   validationMessage?: React.ReactNode | undefined
 }
 
-export const Select: React.FC<SelectProps> & SubComponentProps = ({
+export const Select = ({
   id,
   label,
   description,
@@ -63,10 +57,12 @@ export const Select: React.FC<SelectProps> & SubComponentProps = ({
   status,
   validationMessage,
   classNameOverride,
-  trigger = triggerProps => <TriggerButton {...triggerProps} ref={buttonRef} />,
+  trigger = (triggerProps): JSX.Element => (
+    <TriggerButton {...triggerProps} ref={buttonRef} />
+  ),
   children,
   ...restProps
-}) => {
+}: SelectProps): JSX.Element => {
   const descriptionId = `${id}-field-message`
   const buttonRef = React.useRef<HTMLButtonElement>(null)
   const invalidStatus = status === "error" ? "invalid" : "valid"
@@ -86,7 +82,7 @@ export const Select: React.FC<SelectProps> & SubComponentProps = ({
   const state = useSelectState(ariaSelectProps)
   const renderChildren = children
     ? children
-    : ({ items }) =>
+    : ({ items }): JSX.Element =>
         items.map((item: Node<SingleItemType>) => (
           <Option key={item.key} item={item} />
         ))