+++ packages/split-button/src/SplitButton/components/DropdownButton/DropdownButton.tsx
@@ -7,11 +7,11 @@ import styles from "./DropdownButton.module.scss"
 
 export type DropdownButtonProps = Omit<BaseButtonProps, "label" | "icon">
 
-export const DropdownButton: React.VFC<DropdownButtonProps> = ({
+export const DropdownButton = ({
   classNameOverride,
   "aria-label": ariaLabel,
   ...restProps
-}) => (
+}: DropdownButtonProps): JSX.Element => (
   <BaseButton
     label={ariaLabel || "Open menu"}
     icon={<Icon icon={chevronDown} role="presentation" />}