+++ draft-packages/hero-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.13](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.2.12...@kaizen/draft-hero-card@2.2.13) (2023-01-04)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
 ## [2.2.12](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.2.11...@kaizen/draft-hero-card@2.2.12) (2022-11-15)
 
 **Note:** Version bump only for package @kaizen/draft-hero-card